product after acceptinf functionality

+++ backend/controllers/product.controller.js
@@ -181,13 +181,24 @@ export const updateProduct = async (req, res) => {
 };
 
 export const deleteProduct = async (req, res) => {
-  const { id } = req.params;
-
+  const { id } = req.params; 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ success: false, message: "Invalid Product Id" });
   }
 
   try {
+    const product = await Product.findById(id);
+    if (!product) {
+      return res.status(404).json({ success: false, message: "Product not found" });
+    }
+
+    await User.findByIdAndUpdate(
+      userId,
+      { $pull: { products: id } }, // Remove the product ID from the user's products array
+      { new: true }
+    );
+
+    const userId = product.userId;
     const deletedProduct = await Product.findByIdAndDelete(id);
     if (!deletedProduct) {
       return res.status(404).json({ success: false, message: "Product not found" });
@@ -216,3 +227,6 @@ export const incrementView = async (req,res)=>{
     res.status(500).json({ success: false, message: "Server error", error: error.message });
   }
 };
+
+
+
